requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0
#django-rest-auth>=0.9
#django-allauth>=0.54